refactor: move ArrayBuilderPtr to array_builder.rs
