+++ tskv/src/reader/array_builder.rs
@@ -1,0 +1,159 @@
+use arrow_array::builder::{ArrayBuilder, BooleanBuilder, PrimitiveBuilder, StringBuilder};
+use arrow_array::types::{
+    Float64Type, Int64Type, TimestampMicrosecondType, TimestampMillisecondType,
+    TimestampNanosecondType, TimestampSecondType, UInt64Type,
+};
+use arrow_array::ArrowPrimitiveType;
+use arrow_schema::TimeUnit;
+use models::field_value::DataType;
+use models::schema::tskv_table_schema::PhysicalCType;
+use models::{PhysicalDType, Timestamp};
+use trace::error;
+
+use crate::error::CommonSnafu;
+use crate::TskvResult;
+
+pub struct ArrayBuilderPtr {
+    pub ptr: Box<dyn ArrayBuilder>,
+    pub column_type: PhysicalCType,
+}
+
+impl ArrayBuilderPtr {
+    pub fn new(ptr: Box<dyn ArrayBuilder>, column_type: PhysicalCType) -> Self {
+        Self { ptr, column_type }
+    }
+
+    #[inline(always)]
+    fn builder<T: ArrowPrimitiveType>(&mut self) -> Option<&mut PrimitiveBuilder<T>> {
+        self.ptr.as_any_mut().downcast_mut::<PrimitiveBuilder<T>>()
+    }
+
+    pub fn append_primitive<T: ArrowPrimitiveType>(&mut self, t: T::Native) {
+        if let Some(b) = self.builder::<T>() {
+            b.append_value(t);
+        } else {
+            error!(
+                "Failed to get primitive-type array builder to insert {:?} array",
+                self.column_type
+            );
+        }
+    }
+
+    pub fn append_timestamp(&mut self, unit: &TimeUnit, timestamp: Timestamp) {
+        match unit {
+            TimeUnit::Second => self.append_primitive::<TimestampSecondType>(timestamp),
+            TimeUnit::Millisecond => self.append_primitive::<TimestampMillisecondType>(timestamp),
+            TimeUnit::Microsecond => self.append_primitive::<TimestampMicrosecondType>(timestamp),
+            TimeUnit::Nanosecond => self.append_primitive::<TimestampNanosecondType>(timestamp),
+        }
+    }
+
+    pub fn append_value(
+        &mut self,
+        value_type: PhysicalDType,
+        value: Option<DataType>,
+        column_name: &str,
+    ) -> TskvResult<()> {
+        match value_type {
+            PhysicalDType::Unknown => {
+                return Err(CommonSnafu {
+                    reason: format!("unknown type of column '{}'", column_name),
+                }
+                .build());
+            }
+            PhysicalDType::String => match value {
+                Some(DataType::Str(_, val)) => {
+                    // Safety
+                    // All val is valid UTF-8 String
+                    let str = unsafe { std::str::from_utf8_unchecked(val.as_slice()) };
+                    self.append_string(str)
+                }
+                _ => self.append_null_string(),
+            },
+            PhysicalDType::Boolean => {
+                if let Some(DataType::Bool(_, val)) = value {
+                    self.append_bool(val);
+                } else {
+                    self.append_null_bool();
+                }
+            }
+            PhysicalDType::Float => {
+                if let Some(DataType::F64(_, val)) = value {
+                    self.append_primitive::<Float64Type>(val);
+                } else {
+                    self.append_primitive_null::<Float64Type>();
+                }
+            }
+            PhysicalDType::Integer => {
+                if let Some(DataType::I64(_, val)) = value {
+                    self.append_primitive::<Int64Type>(val);
+                } else {
+                    self.append_primitive_null::<Int64Type>();
+                }
+            }
+            PhysicalDType::Unsigned => {
+                if let Some(DataType::U64(_, val)) = value {
+                    self.append_primitive::<UInt64Type>(val);
+                } else {
+                    self.append_primitive_null::<UInt64Type>();
+                }
+            }
+        }
+        Ok(())
+    }
+
+    pub fn append_primitive_null<T: ArrowPrimitiveType>(&mut self) {
+        if let Some(b) = self.builder::<T>() {
+            b.append_null();
+        } else {
+            error!(
+                "Failed to get primitive-type array builder to insert {:?} array",
+                self.column_type
+            );
+        }
+    }
+
+    pub fn append_bool(&mut self, data: bool) {
+        if let Some(b) = self.ptr.as_any_mut().downcast_mut::<BooleanBuilder>() {
+            b.append_value(data);
+        } else {
+            error!(
+                "Failed to get boolean array builder to insert {:?} array",
+                self.column_type
+            );
+        }
+    }
+
+    pub fn append_null_bool(&mut self) {
+        if let Some(b) = self.ptr.as_any_mut().downcast_mut::<BooleanBuilder>() {
+            b.append_null();
+        } else {
+            error!(
+                "Failed to get boolean array builder to insert {:?} array",
+                self.column_type
+            );
+        }
+    }
+
+    pub fn append_string(&mut self, data: &str) {
+        if let Some(b) = self.ptr.as_any_mut().downcast_mut::<StringBuilder>() {
+            b.append_value(data);
+        } else {
+            error!(
+                "Failed to get string array builder to insert {:?} array",
+                self.column_type
+            );
+        }
+    }
+
+    pub fn append_null_string(&mut self) {
+        if let Some(b) = self.ptr.as_any_mut().downcast_mut::<StringBuilder>() {
+            b.append_null();
+        } else {
+            error!(
+                "Failed to get string array builder to insert {:?} array",
+                self.column_type
+            );
+        }
+    }
+}
